+++ testapp/src/main/java/testapp/endpoint/ghissues/gh554/MsgTemplate.java
@@ -1,0 +1,14 @@
+package testapp.endpoint.ghissues.gh554;
+
+import act.data.annotation.Data;
+import act.util.SimpleBean;
+import org.osgl.util.S;
+
+@Data
+public class MsgTemplate implements SimpleBean {
+    public String id;
+
+    public MsgTemplate(String id) {
+        this.id = id;
+    }
+}